🖥 Setting Revalidation | v1.7.0 🖥
In this revalidation commit, we
updates setting log files from the
official watermark setting to have
an unofficial watermark to show
that the user is using an unofficial
build of NextLanguage.

------------------------------------

With the full files displayed on
the source code. With no hidden
files on the downloaded release.

-------------------------------------

This is an better maintained
Repositories of NextLanguage.

Forked from the official
Repositories from Nxium's
Official NextLanguage's Repo